Update to Bevy 0.17
Currently requires a custom branch with a fix for entity reservation:
https://github.com/bevyengine/bevy/pull/21401
It was already merged and should be available in 0.17.3.

During migration, I noticed that `ScoreboardUi` is not used, so I removed it.

I also disabled tests and benches for the crate since they requires
std and cause rust-analyzer to complain.